requirements.txt
aioredis==1.3.0
amqp==2.5.2
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
beautifulsoup4==4.8.1
billiard==3.6.1.0
celery==4.3.0
cffi==1.13.2
channels==2.3.1
channels-redis==2.4.1
constantly==15.1.0
cryptography==2.8
daphne==2.4.0
Django==2.2.7
get==2019.4.13
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
kombu==4.6.6
more-itertools==7.2.0
msgpack==0.6.2
Pillow==6.2.1
pkg-resources==0.0.0
post==2019.4.13
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyHamcrest==1.9.0
pyOpenSSL==19.1.0
pytz==2019.3
query-string==2019.4.13
redis==3.3.11
request==2019.4.13
service-identity==18.1.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
Twisted==19.10.0
txaio==18.8.1
vine==1.3.0
zipp==0.6.0
zope.interface==4.7.1
